requirements.txt
streamlit
sentence-transformers
faiss-cpu
spacy
torch
numpy
pandas
pypdf
google-generativeai
rank_bm25
transformers
sentencepiece
accelerate
streamlit-option-menu